% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSDF.ERA4v.R
\name{ERA4v.hourly.get.file.name}
\alias{ERA4v.hourly.get.file.name}
\title{ERA4v get file name (hourly)}
\usage{
ERA4v.hourly.get.file.name(variable, year, month, day, hour, stream = "oper",
  fc.init = NULL)
}
\arguments{
\item{variable}{'prmsl', 'prate', 'air.2m', 'uwnd.10m' or 'vwnd.10m' - or any supported variable}

\item{stream}{- 'oper', 'eda' or other ERA5 stream.}

\item{fc.init}{- hour at which the forecast run was initialised (9 or 21) - if NULL
default, uses whicever gives the shortest lead time (ignored for analysis variables).}
}
\value{
File containing the requested data
}
\description{
Get the file name for selected variable and date (hourly data)
}
\details{
Called internally by \code{ERA4v.get.slice.at.hour} but also useful
 called directly - you can then access the data with another tool.
}

